Feature: - Relocated method responsible for Subject age to AnswerSession entity
