+++ src/test/java/com/educator/unit/core/answer_session/AnswerSessionTest.java
@@ -1,8 +1,7 @@
-package com.educator.unit.core.subject;
+package com.educator.unit.core.answer_session;
 
 import com.educator.core.answer_session.AnswerSession;
 import com.educator.core.answer_session.enums.StatusAnswerSession;
-import com.educator.core.subject.SubjectAgeService;
 import com.educator.core.subject.SubjectCompletedAge;
 import org.junit.jupiter.api.Test;
 
@@ -10,15 +9,15 @@ import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class SubjectAgeServiceTest {
+class AnswerSessionTest {
+
     @Test
     void afterSevenDays() {
         //Given
         AnswerSession answerSession1 = new AnswerSession(14L, 10, 7,
                 LocalDate.now().minusDays(8), StatusAnswerSession.COMPLETED, null, null);
         //When
-        SubjectAgeService subjectAgeService = new SubjectAgeService();
-        SubjectCompletedAge result = subjectAgeService.informAboutSubjectAge(answerSession1);
+        SubjectCompletedAge result = answerSession1.getCompletedAge(LocalDate.now());
         //Then
         assertEquals(result, SubjectCompletedAge.OLD);
     }
@@ -29,8 +28,7 @@ class SubjectAgeServiceTest {
         AnswerSession answerSession1 = new AnswerSession(14L, 10, 7,
                 LocalDate.now().minusDays(7), StatusAnswerSession.COMPLETED, null, null);
         //When
-        SubjectAgeService subjectAgeService = new SubjectAgeService();
-        SubjectCompletedAge result = subjectAgeService.informAboutSubjectAge(answerSession1);
+        SubjectCompletedAge result = answerSession1.getCompletedAge(LocalDate.now());
         //Then
         assertEquals(result, SubjectCompletedAge.FRESH);
     }
